Move building of menu data structures into a concern, so menus can work in all areas of main site
